requirements.txt
alembic==1.13.1
altgraph==0.17.4
antiorm==1.2.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
bcrypt==4.3.0
blinker==1.8.0
certifi==2023.11.17
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
crayons==0.4.0
cryptography==44.0.2
db==0.1.1
distlib==0.3.9
filelock==3.16.1
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.0.3
idna==3.6
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.3
Mako==1.3.3
MarkupSafe==2.1.5
mysql-connector-python==8.3.0
mysqlclient==2.2.4
numpy==2.2.2
opencv-python==4.11.0.86
packaging==24.0
pefile==2023.2.7
piexif==1.1.3
pillow==10.4.0
platformdirs==4.3.6
pluggy==1.4.0
psycopg2-binary==2.9.10
pycparser==2.22
pycryptodome==3.21.0
pygame==2.5.2
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
PyMySQL==1.1.0
pytest==8.1.1
python-dotenv==1.1.1
pywin32-ctypes==0.2.3
requests==2.31.0
setuptools==75.8.0
SQLAlchemy==2.0.29
stegano==0.11.4
steganography==0.1.1
typing_extensions==4.11.0
urllib3==2.1.0
virtualenv==20.27.1
Werkzeug==3.0.2
WTForms==3.2.1
